refactor(tool): split _shared helpers by concern
- _shared/helpers.py was a 572-line catch-all and the `build` verb named both response-parsing and payload-building
- split into pagination/fields/custom_fields/parse, rename reads parse_*, dedupe list tail into build_page
