+++ src/mcp_server_polarion/tools/enum.py
@@ -19,13 +19,14 @@ from mcp_server_polarion.core.exceptions import (
 from mcp_server_polarion.models import EnumOption, PaginatedResult
 from mcp_server_polarion.server import mcp
 from mcp_server_polarion.tools._shared.helpers import (
-    DEFAULT_PAGE_SIZE,
-    build_enum_option,
-    compute_has_more,
     encode_path_segment,
-    extract_total_count,
     get_client,
 )
+from mcp_server_polarion.tools._shared.pagination import (
+    DEFAULT_PAGE_SIZE,
+    build_page,
+)
+from mcp_server_polarion.tools._shared.parse import parse_enum_option
 
 
 async def _list_enum_options(  # noqa: PLR0913
@@ -72,22 +73,9 @@ async def _list_enum_options(  # noqa: PLR0913
     if isinstance(data, list):
         for entry in data:
             if isinstance(entry, dict):
-                items.append(build_enum_option(entry))
+                items.append(parse_enum_option(entry))
 
-    raw_total = extract_total_count(response)
-    total = raw_total
-    if total <= 0 and items:
-        total = (page_number - 1) * page_size + len(items)
-
-    return PaginatedResult[EnumOption](
-        items=items,
-        total_count=total,
-        page=page_number,
-        page_size=page_size,
-        has_more=compute_has_more(
-            response, raw_total, page_number, page_size, len(items)
-        ),
-    )
+    return build_page(items, response, page_number, page_size)
 
 
 @mcp.tool(
